Move DeprecationHelper from Utility to Component namespace
