+++ tests/src/Rector/Deprecation/UserPasswordRector/fixture/user_password.php.inc
@@ -11,18 +11,18 @@ function simple_example() {
 <?php
 
 function simple_example() {
-    $password = \Drupal\Core\Utility\DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '9.1.0', function () {
+    $password = \Drupal\Component\Utility\DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '9.1.0', function () {
         return user_password();
     }, function () {
         return \Drupal::service('password_generator')->generate();
     });
-    $other_password = \Drupal\Core\Utility\DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '9.1.0', function () {
+    $other_password = \Drupal\Component\Utility\DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '9.1.0', function () {
         return user_password(8);
     }, function () {
         return \Drupal::service('password_generator')->generate(8);
     });
     $password_length = 12;
-    $last_password = \Drupal\Core\Utility\DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '9.1.0', function () use ($password_length) {
+    $last_password = \Drupal\Component\Utility\DeprecationHelper::backwardsCompatibleCall(\Drupal::VERSION, '9.1.0', function () use ($password_length) {
         return user_password($password_length);
     }, function () use ($password_length) {
         return \Drupal::service('password_generator')->generate($password_length);
